MGDAPI-4638 address feedback @MStokluska
