+++ Makefile
@@ -28,7 +28,7 @@ SMTP_ADDRESS ?= ''
 SMTP_PASS ?= ''
 SMTP_PORT ?= ''
 SMTP_FROM ?= ''
-ROLE_ARN ?= "arn:aws:iam::485026278258:role/12345"
+ROLE_ARN ?= "arn:aws:iam::111111111111:role/11111"
 S3_ACCESS_KEY_ID ?= "123"
 S3_SECRET_ACCESS_KEY ?= "secret"
 TYPE_OF_MANIFEST ?= master
@@ -395,7 +395,7 @@ cluster/prepare/dms:
 
 .PHONY: cluster/prepare/addon-params
 cluster/prepare/addon-params:
-	@-oc process -n $(NAMESPACE) QUOTA=$(DEV_QUOTA) DOMAIN=$(CUSTOM_DOMAIN) StsRoleARN=$(ROLE_ARN) \
+	@-oc process -n $(NAMESPACE) QUOTA=$(DEV_QUOTA) DOMAIN=$(CUSTOM_DOMAIN) STS_ROLE_ARN=$(ROLE_ARN) \
 		S3_ACCESS_KEY_ID=$(S3_ACCESS_KEY_ID) S3_SECRET_ACCESS_KEY=$(S3_SECRET_ACCESS_KEY) \
  		USERNAME=$(SMTP_USER) HOST=$(SMTP_ADDRESS) PASSWORD=$(SMTP_PASS) PORT=$(SMTP_PORT) FROM=$(SMTP_FROM) -f config/secrets/addon-params-secret.yaml | oc apply -f -
