+++ config/secrets/addon-params-secret.yaml
@@ -15,7 +15,7 @@ objects:
       custom-smtp-password: ${PASSWORD}
       custom-smtp-port: ${PORT}
       custom-smtp-from_address: ${FROM}
-      sts-role-arn: ${StsRoleARN}
+      sts-role-arn: ${STS_ROLE_ARN}
       s3-access-key-id: ${S3_ACCESS_KEY_ID}
       s3-secret-access-key: ${S3_SECRET_ACCESS_KEY}
 parameters:
@@ -34,8 +34,8 @@ parameters:
     value: "567"
   - name: FROM
     value: "test@example.com"
-  - name: StsRoleARN
-    value: "arn:aws:iam::485026278258:role/12345"
+  - name: STS_ROLE_ARN
+    value: "arn:aws:iam::111111111111:role/11111"
   - name: S3_ACCESS_KEY_ID
     value: "123"
   - name: S3_SECRET_ACCESS_KEY
