+++ controllers/rhmi/rhmi_controller_test.go
@@ -2,18 +2,14 @@ package controllers
 
 import (
 	"context"
-	"fmt"
 	"reflect"
 	"testing"
 
 	rhmiv1alpha1 "github.com/integr8ly/integreatly-operator/apis/v1alpha1"
-	moqclient "github.com/integr8ly/integreatly-operator/pkg/client"
 	"github.com/integr8ly/integreatly-operator/pkg/config"
 	"github.com/integr8ly/integreatly-operator/pkg/resources"
 	"github.com/integr8ly/integreatly-operator/pkg/resources/marketplace"
-	"github.com/integr8ly/integreatly-operator/pkg/resources/sts"
 	routev1 "github.com/openshift/api/route/v1"
-	olmv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
 	prometheusv1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	"github.com/prometheus/common/model"
 	corev1 "k8s.io/api/core/v1"
@@ -118,106 +114,6 @@ func TestRHMIReconciler_getAlertingNamespace(t *testing.T) {
 				t.Errorf("getAlertingNamespace() got = %v, want %v", got, tt.want)
 			}
 		})
-	}
-}
-
-func Test_validateAddOnStsRoleArnParameterPattern(t *testing.T) {
-	scheme := runtime.NewScheme()
-	_ = corev1.SchemeBuilder.AddToScheme(scheme)
-	_ = olmv1alpha1.SchemeBuilder.AddToScheme(scheme)
-
-	const namespace = "test"
-
-	type args struct {
-		client    client.Client
-		namespace string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    bool
-		wantErr bool
-	}{
-		{
-			name: "test: can't get secret",
-			args: args{
-				client: &moqclient.SigsClientInterfaceMock{
-					ListFunc: func(ctx context.Context, list runtime.Object, opts ...client.ListOption) error {
-						return fmt.Errorf("listError")
-					},
-				},
-				namespace: namespace,
-			},
-			wantErr: true,
-			want:    false,
-		},
-		{
-			name: "test: role arn not found",
-			args: args{
-				client:    fakeclient.NewFakeClientWithScheme(scheme),
-				namespace: namespace,
-			},
-			wantErr: true,
-			want:    false,
-		},
-		{
-			name: "test: role arn empty",
-			args: args{
-				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{sts.RoleArnParameterName: []byte("")})),
-				namespace: namespace,
-			},
-			wantErr: true,
-			want:    false,
-		},
-		{
-			name: "test: role arn regex not match",
-			args: args{
-				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{sts.RoleArnParameterName: []byte("notAnARN")})),
-				namespace: namespace,
-			},
-			wantErr: true,
-			want:    false,
-		},
-		{
-			name: "test: role arn regex match",
-			args: args{
-				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{sts.RoleArnParameterName: []byte("arn:aws:iam::123456789012:role/12345")})),
-				namespace: namespace,
-			},
-			wantErr: false,
-			want:    true,
-		},
-		{
-			name: "test: role arn regex match for AWS GovCloud (US) Regions",
-			args: args{
-				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{sts.RoleArnParameterName: []byte("arn:aws-us-gov:iam::123456789012:role/12345")})),
-				namespace: namespace,
-			},
-			wantErr: false,
-			want:    true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := validateAddOnStsRoleArnParameterPattern(tt.args.client, tt.args.namespace)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("validateAddOnStsRoleArnParameterPattern() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("validateAddOnStsRoleArnParameterPattern() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func buildAddonSecret(namespace string, secretData map[string][]byte) *corev1.Secret {
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "addon-managed-api-service-parameters",
-			Namespace: namespace,
-		},
-		Data: secretData,
 	}
 }
