+++ pkg/products/cloudresources/reconciler_test.go
@@ -2,7 +2,6 @@ package cloudresources
 
 import (
 	"context"
-	"github.com/integr8ly/integreatly-operator/pkg/resources/sts"
 	"testing"
 
 	threescalev1 "github.com/3scale/3scale-operator/apis/apps/v1alpha1"
@@ -25,7 +24,6 @@ import (
 	operatorsv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	apiextensionv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -245,90 +243,4 @@ func getBuildScheme() (*runtime.Scheme, error) {
 	}
 
 	return scheme, err
-}
-
-func TestReconciler_createSTSArnSecret(t *testing.T) {
-	scheme, err := getBuildScheme()
-	if err != nil {
-		t.Fatalf("Error obtaining scheme")
-	}
-
-	type fields struct {
-		Config        *config.CloudResources
-		ConfigManager config.ConfigReadWriter
-		installation  *integreatlyv1alpha1.RHMI
-		mpm           marketplace.MarketplaceInterface
-		log           logger.Logger
-		Reconciler    *resources.Reconciler
-		recorder      record.EventRecorder
-	}
-	type args struct {
-		ctx               context.Context
-		client            client.Client
-		operatorNamespace string
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		want    integreatlyv1alpha1.StatusPhase
-		wantErr bool
-	}{
-		{
-			name: "test: phase failed on error getting role arn",
-			fields: fields{
-				log: getLogger(),
-				installation: &integreatlyv1alpha1.RHMI{
-					ObjectMeta: metav1.ObjectMeta{Namespace: defaultInstallationNamespace},
-				},
-			},
-			args: args{
-				client: moqclient.NewSigsClientMoqWithScheme(scheme),
-			},
-			want:    integreatlyv1alpha1.PhaseFailed,
-			wantErr: true,
-		},
-		{
-			name: "test: phase complete on creating secret",
-			fields: fields{
-				log: getLogger(),
-				installation: &integreatlyv1alpha1.RHMI{
-					ObjectMeta: metav1.ObjectMeta{Namespace: defaultInstallationNamespace},
-				},
-			},
-			args: args{
-				client: moqclient.NewSigsClientMoqWithScheme(scheme, &corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "addon-managed-api-service-parameters",
-						Namespace: defaultInstallationNamespace,
-					},
-					Data: map[string][]byte{
-						sts.RoleArnParameterName: []byte("arn:aws:iam::123456789012:role/12345"),
-					},
-				}),
-			},
-			want: integreatlyv1alpha1.PhaseCompleted,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &Reconciler{
-				Config:        tt.fields.Config,
-				ConfigManager: tt.fields.ConfigManager,
-				installation:  tt.fields.installation,
-				mpm:           tt.fields.mpm,
-				log:           tt.fields.log,
-				Reconciler:    tt.fields.Reconciler,
-				recorder:      tt.fields.recorder,
-			}
-			got, err := r.createSTSARNSecret(tt.args.ctx, tt.args.client, tt.args.operatorNamespace)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("createSTSARNSecret() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("createSTSARNSecret() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
