+++ pkg/resources/sts/sts_test.go
@@ -3,9 +3,11 @@ package sts
 import (
 	"context"
 	"fmt"
+	integreatlyv1alpha1 "github.com/integr8ly/integreatly-operator/apis/v1alpha1"
 	moqclient "github.com/integr8ly/integreatly-operator/pkg/client"
 	"github.com/integr8ly/integreatly-operator/pkg/resources/logger"
 	cloudcredentialv1 "github.com/openshift/api/operator/v1"
+	olmv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -22,6 +24,10 @@ func getBuildScheme() (*runtime.Scheme, error) {
 		return nil, err
 	}
 	err = corev1.AddToScheme(scheme)
+	if err != nil {
+		return nil, err
+	}
+	err = olmv1alpha1.AddToScheme(scheme)
 	if err != nil {
 		return nil, err
 	}
@@ -102,6 +108,164 @@ func TestIsClusterSTS(t *testing.T) {
 			}
 			if got != tt.want {
 				t.Errorf("IsClusterSTS() got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_ValidateAddOnStsRoleArnParameterPattern(t *testing.T) {
+	scheme := runtime.NewScheme()
+	_ = corev1.SchemeBuilder.AddToScheme(scheme)
+	_ = olmv1alpha1.SchemeBuilder.AddToScheme(scheme)
+
+	const namespace = "test"
+
+	type args struct {
+		client    client.Client
+		namespace string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    bool
+		wantErr bool
+	}{
+		{
+			name: "test: can't get secret",
+			args: args{
+				client: &moqclient.SigsClientInterfaceMock{
+					ListFunc: func(ctx context.Context, list runtime.Object, opts ...client.ListOption) error {
+						return fmt.Errorf("listError")
+					},
+				},
+				namespace: namespace,
+			},
+			wantErr: true,
+			want:    false,
+		},
+		{
+			name: "test: role arn not found",
+			args: args{
+				client:    fakeclient.NewFakeClientWithScheme(scheme),
+				namespace: namespace,
+			},
+			wantErr: true,
+			want:    false,
+		},
+		{
+			name: "test: role arn empty",
+			args: args{
+				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{RoleArnParameterName: []byte("")})),
+				namespace: namespace,
+			},
+			wantErr: true,
+			want:    false,
+		},
+		{
+			name: "test: role arn regex not match",
+			args: args{
+				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{RoleArnParameterName: []byte("notAnARN")})),
+				namespace: namespace,
+			},
+			wantErr: true,
+			want:    false,
+		},
+		{
+			name: "test: role arn regex match",
+			args: args{
+				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{RoleArnParameterName: []byte("arn:aws:iam::123456789012:role/12345")})),
+				namespace: namespace,
+			},
+			wantErr: false,
+			want:    true,
+		},
+		{
+			name: "test: role arn regex match for AWS GovCloud (US) Regions",
+			args: args{
+				client:    fakeclient.NewFakeClientWithScheme(scheme, buildAddonSecret(namespace, map[string][]byte{RoleArnParameterName: []byte("arn:aws-us-gov:iam::123456789012:role/12345")})),
+				namespace: namespace,
+			},
+			wantErr: false,
+			want:    true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ValidateAddOnStsRoleArnParameterPattern(tt.args.client, tt.args.namespace)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("validateAddOnStsRoleArnParameterPattern() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got != tt.want {
+				t.Errorf("validateAddOnStsRoleArnParameterPattern() got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func buildAddonSecret(namespace string, secretData map[string][]byte) *corev1.Secret {
+	return &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "addon-managed-api-service-parameters",
+			Namespace: namespace,
+		},
+		Data: secretData,
+	}
+}
+
+func Test_CreateSTSArnSecret(t *testing.T) {
+	scheme, err := getBuildScheme()
+	if err != nil {
+		t.Fatalf("Error obtaining scheme")
+	}
+
+	type args struct {
+		ctx                   context.Context
+		client                client.Client
+		installationNamespace string
+		operatorNamespace     string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    integreatlyv1alpha1.StatusPhase
+		wantErr bool
+	}{
+		{
+			name: "test: phase failed on error getting role arn",
+			args: args{
+				client: moqclient.NewSigsClientMoqWithScheme(scheme),
+			},
+			want:    integreatlyv1alpha1.PhaseFailed,
+			wantErr: true,
+		},
+		{
+			name: "test: phase complete on creating secret",
+			args: args{
+				client: moqclient.NewSigsClientMoqWithScheme(scheme, &corev1.Secret{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "addon-managed-api-service-parameters",
+						Namespace: "redhat-rhoam-operator",
+					},
+					Data: map[string][]byte{
+						RoleArnParameterName: []byte("arn:aws:iam::123456789012:role/12345"),
+					},
+				}),
+				installationNamespace: "redhat-rhoam-operator",
+				operatorNamespace:     "",
+			},
+			want: integreatlyv1alpha1.PhaseCompleted,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := CreateSTSARNSecret(tt.args.ctx, tt.args.client, tt.args.installationNamespace, tt.args.operatorNamespace)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("CreateSTSARNSecret() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got != tt.want {
+				t.Errorf("createSTSARNSecret() got = %v, want %v", got, tt.want)
 			}
 		})
 	}
